requirements.txt
#fastapi
#uvicorn
#python-multipart
#thefuzz
#flask

flask
flask-cors
google-generativeai
python-dotenv
gunicorn